requirements.txt
streamlit
google-generativeai
PyPDF2
sentence-transformers
faiss-cpu
scikit-learn
rouge
python-dotenv
